+++ app/src/main/java/com/example/software_engineering_project/controller/FragmentCleaningPlanAddController.java
@@ -1,0 +1,26 @@
+package com.example.software_engineering_project.controller;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.software_engineering_project.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link FragmentCleaningPlanAddController #newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class FragmentCleaningPlanAddController extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_cleaning_plan_add, container, false);
+    }
+}
