+++ app/src/main/java/com/example/software_engineering_project/controller/FragmentCleaningPlanController.java
@@ -1,0 +1,52 @@
+package com.example.software_engineering_project.controller;
+
+import android.content.Context;
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
+import android.widget.Toast;
+
+import com.example.software_engineering_project.R;
+import com.example.software_engineering_project.adapter.CleaningPlanListViewAdapter;
+
+import java.util.ArrayList;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link FragmentCleaningPlanController #newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class FragmentCleaningPlanController extends Fragment {
+
+    View fragmentView;
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+
+
+        FragmentCleaningPlanListController fragmentCleaningPlanListController = new FragmentCleaningPlanListController();
+        callFragment(fragmentCleaningPlanListController);
+        fragmentView = inflater.inflate(R.layout.fragment_cleaning_plan, container, false);
+        //context = getActivity();
+        return fragmentView;
+    }
+
+
+
+    private void callFragment(Fragment fragment){
+
+        FragmentManager fm = getActivity().getSupportFragmentManager();
+        FragmentTransaction transaction = fm.beginTransaction();
+        transaction.replace(R.id.contentFragmentCleaningPlan, fragment);
+        transaction.commit();
+
+    }
+}
