+++ app/src/main/java/com/example/software_engineering_project/controller/FragmentManageFlatShareController.java
@@ -1,10 +1,11 @@
 package com.example.software_engineering_project.controller;
 
-import androidx.appcompat.app.AppCompatActivity;
 import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
+import androidx.fragment.app.Fragment;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
@@ -16,78 +17,91 @@ import java.io.FileOutputStream;
 import java.util.ArrayList;
 
 import com.example.software_engineering_project.R;
-import com.example.software_engineering_project.adapter.ManageFlatShareListViewAdapter;
+import com.example.software_engineering_project.adapter.CreateFlatShareListViewAdapter;
 
-public class ManageFlatShareController extends AppCompatActivity {
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link FragmentManageFlatShareController #newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class FragmentManageFlatShareController extends Fragment {
 
+    View fragmentView;
     static ListView listView;
-    EditText input;
+    EditText input_name, input_mail;
     ImageView enter;
-    static ManageFlatShareListViewAdapter adapter;
-    static ArrayList<String> items;
+    static CreateFlatShareListViewAdapter adapter;
+    static ArrayList<String> items = new ArrayList<>();
     static Context context;
-    private Button goBackButtonManageFlatShare;
+    private Button cancelCreateFlatShare;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.fragment_manage_flat_share_screen);
-
-        this.addButtons();
-
-        listView = findViewById(R.id.groceryList);
-        input = findViewById(R.id.input);
-        enter = findViewById(R.id.add);
-        context = getApplicationContext();
-
-        // add hardcoded items to list
-        items = new ArrayList<>();
-
-        //Auslesebefehl für Datenbank
-        items.add("Jonas");
-        items.add("Nikos");
-        items.add("Lucas");
+        adapter = new CreateFlatShareListViewAdapter(getActivity(), items);
         items.add("Meike");
+        items.add("Lucas");
         items.add("Laura");
+        items.add("Nikos");
+        items.add("Jonas");
+
+        fragmentView = inflater.inflate(R.layout.fragment_manage_flat_share, container, false);
+
+        context = getActivity();
+        listView = fragmentView.findViewById(R.id.flatShareMemberList);
+        input_name = fragmentView.findViewById(R.id.input_name);
+        input_mail = fragmentView.findViewById(R.id.input_mail);
+        enter = fragmentView.findViewById(R.id.addFlatShareMember);
 
         listView.setLongClickable(true);
-        adapter = new ManageFlatShareListViewAdapter(this, items);
         listView.setAdapter(adapter);
 
-        // Display the person's name when the person's row is clicked
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 String clickedItem = (String) listView.getItemAtPosition(position);
-                Toast.makeText(ManageFlatShareController.this, clickedItem, Toast.LENGTH_SHORT).show();
+                Toast.makeText(getActivity(), clickedItem, Toast.LENGTH_SHORT).show();
             }
 
         });
 
-        // add person when the user presses the add button
-        enter.setOnClickListener(new View.OnClickListener() {
+        listView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
+
+            @Override
+            public boolean onItemLongClick(AdapterView<?> adapterView, View view, int position, long l) {
+                removeItem(position);
+                return false;
+            }
+
+        });
+
+        enter.setOnClickListener(new AdapterView.OnClickListener() {
+
             @Override
             public void onClick(View view) {
-                String text = input.getText().toString();
+                String text = input_mail.getText().toString();
                 if (text.length() == 0) {
-                    makeToast("Enter an e-mail.");
+                    makeToast("Enter an e-mail");
                 } else {
                     addItem(text);
-                    input.setText("");
+                    input_mail.setText("");
                     makeToast("Added " + text);
                 }
+
             }
 
         });
+
+        return fragmentView;
 
     }
 
     // Override onDestroy() to save the contents of the grocery list right before the app is terminated
     @Override
-    protected void onDestroy() {
+    public void onDestroy() {
 
-        File path = getApplicationContext().getFilesDir();
+        File path = context.getApplicationContext().getFilesDir();
         try {
             FileOutputStream writer = new FileOutputStream(new File(path, "list.txt"));
             writer.write(items.toString().getBytes());
@@ -99,13 +113,17 @@ public class ManageFlatShareController extends AppCompatActivity {
 
     }
 
-    // function to remove a member given its index in the list.
+    // function to remove an item given its index in the grocery list.
     public static void removeItem(int i) {
 
         makeToast("Removed: " + items.get(i));
         items.remove(i);
         listView.setAdapter(adapter);
 
+    }
+
+    public static void uncheckItem(int i) {
+        makeToast("Unchecked: " + items.get(i));
     }
 
     // function to add an item given its name.
@@ -124,16 +142,6 @@ public class ManageFlatShareController extends AppCompatActivity {
         if (t != null) t.cancel();
         t = Toast.makeText(context, s, Toast.LENGTH_SHORT);
         t.show();
-
-    }
-
-    private void addButtons() {
-
-        goBackButtonManageFlatShare = findViewById(R.id.goBackButtonManageFlatShare);
-        goBackButtonManageFlatShare.setOnClickListener(view -> {
-            Intent SettingScreen = new Intent(ManageFlatShareController.this, FragmentSettingsController.class);
-            startActivity(SettingScreen);
-        });
 
     }
