+++ app/src/main/res/drawable-anydpi/add_symbol_gray.xml
