--- app/src/main/res/layout/activity_mangage_flat_share_list_view_adapter.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="15sp">
-
-    <TextView
-        android:id="@+id/number1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignBaseline="@id/name1"
-        android:layout_marginBottom="2sp"
-        android:textColor="#A9E0F4"
-        android:textStyle="bold" />
-
-    <TextView
-        android:id="@+id/name1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_marginLeft="28sp"
-        android:textColor="#000"
-        android:textSize="20sp" />
-<!--// maybe for future release-->
-
-<!--    <ImageView
-        android:id="@+id/edit"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_marginRight="16dp"
-        android:layout_toLeftOf="@id/remove"
-        android:src="@drawable/edit" />-->
-
-    <ImageView
-        android:id="@+id/remove1"
-        android:layout_width="22sp"
-        android:layout_height="22sp"
-        android:layout_alignParentRight="true"
-        android:layout_centerVertical="true"
-        android:layout_marginRight="18sp"
-        android:src="@drawable/remove" />
-
-</RelativeLayout>
