+++ app/src/main/res/layout/manage_flat_share_list_view_adapter.xml
